refactor: remove RDS synthetic suite re-export shims
