+++ tests/synthetic/rds_postgres/scoring.py
@@ -2,9 +2,6 @@
 
 This module is free of ``app.*`` imports so scoring logic can be unit-tested
 without importing the full investigation runtime or any heavy runtime dependencies.
-
-Dataclasses and functions in this module are re-exported from run_suite.py
-for backward compatibility with existing import sites.
 """
 
 from __future__ import annotations
